+++ packages/components/psammead-bulletin/src/index.stories.jsx
@@ -7,13 +7,20 @@ import notes from '../README.md';
 import Bulletin from '.';
 
 /* eslint-disable react/prop-types */
-const BulletinComponent = ({ script, service, type, hasImage, dir, text }) => {
+const BulletinComponent = ({
+  script,
+  service,
+  mediaType,
+  hasImage,
+  dir,
+  text,
+}) => {
   const ctaLink = 'https://bbc.co.uk';
 
   const isLive = boolean('Live', false);
-  const ctaText = type === 'audio' ? 'Listen' : 'Watch';
+  const ctaText = mediaType === 'audio' ? 'Listen' : 'Watch';
+  const offScreenText = isLive ? `${ctaText} Live` : ctaText;
 
-  const liveCtaText = isLive ? `${ctaText} Live` : ctaText;
   const image = (
     <Image
       src="https://ichef.bbci.co.uk/news/660/cpsprodpb/11897/production/_106613817_999_al_.jpg"
@@ -24,15 +31,16 @@ const BulletinComponent = ({ script, service, type, hasImage, dir, text }) => {
   return (
     <Bulletin
       image={hasImage && image}
-      type={type}
+      mediaType={mediaType}
       isLive={isLive}
       script={script}
       service={service}
       headlineText={text}
       summaryText={text}
       ctaLink={ctaLink}
-      ctaText={liveCtaText}
+      ctaText={ctaText}
       dir={dir}
+      offScreenText={offScreenText}
     />
   );
 };
@@ -46,7 +54,7 @@ storiesOf('Components|Bulletin', module)
       <BulletinComponent
         script={script}
         service={service}
-        type="video"
+        mediaType="video"
         hasImage
         dir={dir}
         text={textSnipet}
@@ -63,7 +71,7 @@ storiesOf('Components|Bulletin', module)
         <BulletinComponent
           script={script}
           service={service}
-          type="audio"
+          mediaType="audio"
           hasImage={hasImage}
           dir={dir}
           text={textSnipet}