+++ packages/components/psammead-bulletin/src/index.test.jsx
@@ -5,10 +5,13 @@ import Image from '@bbc/psammead-image';
 import Bulletin from '.';
 
 /* eslint-disable react/prop-types */
-const BulletinComponent = ({ script, service, isLive, type, ctaText }) => {
+const BulletinComponent = ({ script, service, isLive, mediaType, ctaText }) => {
   const summaryText = 'This is the summary text';
   const headlineText = 'This is the headline';
   const ctaLink = 'https://bbc.co.uk';
+
+  const playCtaText = isLive ? `${ctaText} Live` : ctaText;
+  const offScreenText = isLive ? `${ctaText} LIVE` : ctaText;
 
   const image = (
     <Image
@@ -19,14 +22,15 @@ const BulletinComponent = ({ script, service, isLive, type, ctaText }) => {
   return (
     <Bulletin
       image={image}
-      type={type}
+      mediaType={mediaType}
       isLive={isLive}
       script={script}
       service={service}
       headlineText={headlineText}
       summaryText={summaryText}
       ctaLink={ctaLink}
-      ctaText={ctaText}
+      ctaText={playCtaText}
+      offScreenText={offScreenText}
     />
   );
 };
@@ -38,7 +42,7 @@ describe('Bulletin', () => {
       script={latin}
       service="news"
       ctaText="Listen"
-      type="audio"
+      mediaType="audio"
     />,
   );
 
@@ -48,7 +52,7 @@ describe('Bulletin', () => {
       script={latin}
       service="news"
       ctaText="Watch"
-      type="video"
+      mediaType="video"
     />,
   );
 
@@ -58,7 +62,7 @@ describe('Bulletin', () => {
       script={latin}
       service="news"
       ctaText="Listen"
-      type="audio"
+      mediaType="audio"
       isLive
     />,
   );
@@ -69,7 +73,7 @@ describe('Bulletin', () => {
       script={latin}
       service="news"
       ctaText="Watch"
-      type="video"
+      mediaType="video"
       isLive
     />,
   );